Move test-support helpers to addon-test-support
