+++ tests/integration/components/helpers-test.js
@@ -2,7 +2,7 @@ import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 import Coordinator from '../../../models/coordinator';
-import { drag } from '../../helpers/drag-drop';
+import { drag } from 'ember-drag-drop/test-support/helpers/drag-drop';
 import { find, render } from '@ember/test-helpers';
 
 module('Integration | Helpers', function(hooks) {
@@ -65,4 +65,3 @@ module('Integration | Helpers', function(hooks) {
   });
 
 });
-
